resourceproviders: polling at 10s intervals to ensure these are registered
