+++ internal/provider/provider.go
@@ -8,6 +8,7 @@ import (
 	"log"
 	"os"
 	"strings"
+	"time"
 
 	"github.com/hashicorp/go-azure-sdk/sdk/auth"
 	"github.com/hashicorp/go-azure-sdk/sdk/environments"
@@ -458,7 +459,10 @@ func buildClient(ctx context.Context, p *schema.Provider, d *schema.ResourceData
 	if !skipProviderRegistration {
 		subscriptionId := commonids.NewSubscriptionID(client.Account.SubscriptionId)
 		requiredResourceProviders := resourceproviders.Required()
-		if err := resourceproviders.EnsureRegistered(ctx, client.Resource.ResourceProvidersClient, subscriptionId, requiredResourceProviders); err != nil {
+		ctx2, cancel := context.WithTimeout(ctx, 30*time.Minute)
+		defer cancel()
+		
+		if err := resourceproviders.EnsureRegistered(ctx2, client.Resource.ResourceProvidersClient, subscriptionId, requiredResourceProviders); err != nil {
 			return nil, diag.Errorf(resourceProviderRegistrationErrorFmt, err)
 		}
 	}
